Corelib now compiles cleanly with /w64 & still runs as a 32-bit app. Untested (yet) as a 64-bit build.
git-svn-id: svn://svn.code.sf.net/p/passwordsafe/code/trunk/pwsafe/pwsafe@1354 1f79f812-37fb-46fe-a122-30589dd2bf55

+++ corelib/ItemData.cpp
@@ -920,7 +920,7 @@ CItemData::ValidatePWHistory()
     return 0;
   }
 
-  int listnum = pPWHistList->GetCount();
+  int listnum = static_cast<int>(pPWHistList->GetCount());
   if (listnum > pwh_num)
     pwh_num = listnum;
